address review: move test to test_initializer.py and simplify release note
