+++ test/python/circuit/test_initializer.py
@@ -523,6 +523,27 @@ class TestInitialize(QiskitTestCase):
             _ = circuit.inverse()
 
 
+    def test_gates_to_uncompute_integer_input(self):
+        """Test that gates_to_uncompute works correctly for integer-bitmap inputs.
+        Regression test for https://github.com/Qiskit/qiskit/issues/16413
+        """
+        # Integer-bitmap input: Initialize(1, num_qubits=2) prepares |01>
+        int_init = Initialize(1, num_qubits=2)
+        uncompute_int = int_init.gates_to_uncompute()
+        self.assertEqual(uncompute_int.num_qubits, 2)
+
+        # Equivalent amplitude-vector input should give same result
+        vec_init = Initialize([0, 1, 0, 0])
+        uncompute_vec = vec_init.gates_to_uncompute()
+        self.assertEqual(uncompute_vec.num_qubits, 2)
+
+        # Verify integer-bitmap uncompute actually maps |01> back to |00...0>
+        from qiskit.quantum_info import Statevector
+        result = Statevector.from_label("01").evolve(uncompute_int)
+        zero_state = Statevector.from_label("00")
+        self.assertTrue(result.equiv(zero_state))
+
+
 class TestInstructionParam(QiskitTestCase):
     """Test conversion of numpy type parameters."""
